iuphd.cls Delete \raggedright command in the definition of abstract page so that it follows the following instruction:
https://graduate.indiana.edu/doc/academic-research/Sample%20Abstract%20tips%20page%204.pdf
Committee names and signature lines are right justified while the student name and title are centered and the text of the abstract is justified as the rest of the dissertation – either left justified or justified on both sides. Behind each committee member’s name, place the appropriate post‐nominal initials for that individual (such as Jane Doe, PhD).